+++ printlinter/classes.py
@@ -3,7 +3,7 @@
 # Standard imports
 from collections import namedtuple
 from dataclasses import dataclass
-from enum import Enum
+from enum import Enum, auto
 from pathlib import Path
 
 Issue = namedtuple("Issue", ["err_code", "name"])
@@ -22,6 +22,15 @@ class IssueEnum(Issue, Enum):
     SYSSTDERRWRITELINESDETECT = Issue(
         err_code="PPL006", name="sys.stderr.writelines-detected"
     )
+
+
+class OutputLevel(Enum):
+    """Minimize level enum."""
+
+    DEFAULT = auto()
+    L1 = auto()
+    L2 = auto()
+    L3 = auto()
 
 
 @dataclass